fix(cli-sub-agent): repo-write audit uses per-execution baseline (#910 round 9)
Round-8 left the audit comparing against git_head_at_creation / pre_session_porcelain captured once at session birth. A reused session where turn 1 modified tracked files and turn 2 resumed without edits would still emit a repo_write_audit warning on turn 2 because the baseline was stale by a whole turn.

Capture a fresh (HEAD, porcelain) snapshot at the start of every execution, thread it through PostExecContext, and prefer it over the session-creation baseline in the audit compute step. Session-creation fields remain as a resilient fallback when per-execution capture fails.

Closes #910.
